easy-pipelines/requirements.txt
flask
google-api-python-client
google-cloud-dataproc
google-cloud-storage